+++ HISTORY.rst
@@ -11,6 +11,7 @@ Next release
 - Add `admin_logs_streaming` support for events stream (`#623 <https://github.com/box/box-python-sdk/pull/623>`_)
 - Add `vanity_name` param for creating shared link to a file or folder (`#637 <https://github.com/box/box-python-sdk/pull/637>`_)
 - Add getting files and file versions under retention for a retention policy assignment (`#633 <https://github.com/box/box-python-sdk/pull/633>`_)
+- Support base item operations for WebLink class (`#639 <https://github.com/box/box-python-sdk/pull/639>`_)
 
 **Bug Fixes:**
 - Limit cryptography to version <3.5.0 (`#636 <https://github.com/box/box-python-sdk/pull/636>`_)